+++ BASE/Test/TestFramework/Shared/LocalInProcHttpServer.cs
@@ -7,6 +7,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.Net.Http.Headers;
 
 namespace Microsoft.ApplicationInsights.TestFramework
 {
@@ -14,8 +15,12 @@ namespace Microsoft.ApplicationInsights.TestFramework
     {
         private readonly IWebHost host;
         private readonly CancellationTokenSource cts;
+        
+        public int RequestCounter = 0;
 
-        public RequestDelegate ServerLogic = async (httpContext) => await httpContext.Response.WriteAsync("Hello World!");
+        public Func<HttpContext, Task> ServerLogic;
+
+        public Action<HttpContext> ServerSideAsserts;
 
         public LocalInProcHttpServer(string url)
         {
@@ -25,11 +30,20 @@ namespace Microsoft.ApplicationInsights.TestFramework
                 .UseUrls(url)
                 .Configure((app) =>
                 {
-                    app.Run(ServerLogic);
+                    app.Run(Server);
                 })
                 .Build();
 
             Task.Run(() => this.host.RunAsync(this.cts.Token));
+        }
+
+        private Task Server(HttpContext context)
+        {
+            Interlocked.Increment(ref this.RequestCounter);
+            
+            this.ServerSideAsserts?.Invoke(context);
+
+            return this.ServerLogic(context);
         }
 
         public void Dispose()
@@ -42,6 +56,36 @@ namespace Microsoft.ApplicationInsights.TestFramework
             catch (Exception)
             {
             }
+        }
+
+        public static LocalInProcHttpServer MakeRedirectServer(string url, string redirectUrl, TimeSpan cacheExpirationDuration)
+        {
+            return new LocalInProcHttpServer(url)
+            {
+                ServerLogic = async (httpContext) =>
+                {
+                    httpContext.Response.StatusCode = StatusCodes.Status308PermanentRedirect;
+                    httpContext.Response.Headers.Add("Location", redirectUrl);
+
+                    // https://docs.microsoft.com/en-us/aspnet/core/performance/caching/middleware?view=aspnetcore-5.0
+                    // https://docs.microsoft.com/en-us/dotnet/api/system.net.http.headers.cachecontrolheadervalue?view=net-5.0
+                    httpContext.Response.GetTypedHeaders().CacheControl = new CacheControlHeaderValue()
+                    {
+                        Public = true,
+                        MaxAge = cacheExpirationDuration,
+                    };
+
+                    await httpContext.Response.WriteAsync("redirect");
+                },
+            };
+        }
+
+        public static LocalInProcHttpServer MakeTargetServer(string url, string response)
+        {
+            return new LocalInProcHttpServer(url)
+            {
+                ServerLogic = async (httpContext) => await httpContext.Response.WriteAsync(response)
+            };
         }
     }
 }