+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - [The `{OriginalFormat}` field in ILogger Scope will be emitted as `OriginalFormat` with the braces removed](https://github.com/microsoft/ApplicationInsights-dotnet/pull/2362)
 - [Enable SDK to create package for auto-instrumentation](https://github.com/microsoft/ApplicationInsights-dotnet/pull/2365)
 - [Suppress long running SignalR connections](https://github.com/microsoft/ApplicationInsights-dotnet/pull/2372)
+- [Support for redirect response from Ingestion service](https://github.com/microsoft/ApplicationInsights-dotnet/issues/2327)
+- [NetCore2.1 has reached end of life and all references have been removed](https://github.com/microsoft/ApplicationInsights-dotnet/issues/2251)
 
 ## Version 2.18.0
 - [Change Self-Diagnostics to include datetimestamp in filename](https://github.com/microsoft/ApplicationInsights-dotnet/pull/2325)